portal_backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pyjwt
email-validator
